requirements.txt
# -------------------------------
# Core
# -------------------------------
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
pillow>=10.0.0


# -------------------------------
# API
# -------------------------------
fastapi==0.111.0
uvicorn==0.30.1

# -------------------------------
# LLM (Qwen-2.5)
# -------------------------------
torch>=2.1.0
transformers>=4.41.0
accelerate>=0.30.0
sentencepiece>=0.2.0
safetensors>=0.4.3

# -------------------------------
# Messaging (since you use NATS)
# -------------------------------
nats-py==2.8.0

# -------------------------------
# Utilities
# -------------------------------
loguru==0.7.2
requests==2.32.3
